refactor(desktop): make the port seam safe to work on in parallel
Closes #261.

Three things made concurrent ports collide. None of them changes a byte
on the wire — every live diff still passes — but together they were the
reason two ports in flight always touched the same hunks.

**1. The route table is a registry, not one match statement.**

Each ported module now declares its own `claims`/`serve` as a
`native::Endpoint`, and `native/mod.rs` walks an `ENDPOINTS` list.
Adding an endpoint is one appended line plus its own file, where it used
to be an edit to a shared `match` covering every route. `mod.rs` knows
nothing about what a module does, and no module knows about another.

The pair travels together on purpose: claiming a route and implementing
it are the same edit, so a route cannot end up claimed by a handler that
does not exist — which fails as a *silent* fallback to Go rather than as
a compile error. `no_two_endpoints_claim_the_same_request` guards the one
failure a registry has that a match did not: two modules claiming one
path, where the first listed wins and the other's tests keep passing.
`every_registered_endpoint_claims_something` catches the reverse — an
entry whose `claims` never fires, which reads as a shipped port.

**2. One live-parity suite per area.**

`tests/live_parity.rs` held every endpoint's diff, so it was the file
every port edited. Split into `parity_pricing`, `parity_sessions`,
`parity_agents` and `parity_analytics`, sharing `tests/parity_common/`.
Each is its own binary, so `cargo test --test parity_analytics` runs one
area rather than all of them.

**3. `parity-instance.sh` is per checkout.**

It hardcoded `/tmp/agento-parity`, so two agents shared one scratch
database, one built binary and one PID file — and `start` begins by
calling `stop`, so the second agent killed the first's server. The work
dir now defaults to a name derived from the checkout's own path, and
`stop` only ever kills the PID recorded in that dir. Two agents in
separate worktrees need no coordination; two sharing a checkout set
`AGENTO_PARITY_WORKER=<id>` or `AGENTO_PARITY_DIR=<path>`.

`start` also exports `AGENTO_PARITY_DIR` alongside the URL, and a new
`url` subcommand re-prints the exports for a shell that lost them without
restarting anything.

Verified: 87 Rust unit tests, all six live-diff cases byte-identical
against a Go server built from this checkout, and two workers started
side by side — both healthy, and stopping one left the other running.
